+++ generators/messaging-adapter-generator/src/test/java/uk/gov/justice/subscription/jms/it/JmsEndpointGenerationCustomIT.java
@@ -9,16 +9,18 @@ import uk.gov.justice.api.subscription.CustomEventListenerExampleEventEventFilte
 import uk.gov.justice.api.subscription.CustomEventListenerExampleEventEventListenerInterceptorChainProvider;
 import uk.gov.justice.api.subscription.CustomEventListenerExampleEventEventValidationInterceptor;
 import uk.gov.justice.api.subscription.CustomEventListenerExampleEventJmsListener;
+import uk.gov.justice.api.subscription.CustomEventListenerExampleEventJmsLoggerMetadataInterceptor;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventFilter;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventFilterInterceptor;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventListenerInterceptorChainProvider;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventValidationInterceptor;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventJmsListener;
+import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventJmsLoggerMetadataInterceptor;
 import uk.gov.justice.schema.service.SchemaCatalogResolverProducer;
 import uk.gov.justice.services.adapter.messaging.DefaultJmsParameterChecker;
 import uk.gov.justice.services.adapter.messaging.DefaultJmsProcessor;
 import uk.gov.justice.services.adapter.messaging.DefaultSubscriptionJmsProcessor;
-import uk.gov.justice.services.adapter.messaging.JmsLoggerMetadataInterceptor;
+import uk.gov.justice.services.adapter.messaging.JmsLoggerMetadataAdder;
 import uk.gov.justice.services.adapter.messaging.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.cdi.LoggerProducer;
 import uk.gov.justice.services.common.configuration.GlobalValueProducer;
@@ -102,12 +104,14 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
             CustomEventListenerPeopleEventEventListenerInterceptorChainProvider.class,
             CustomEventListenerPeopleEventEventValidationInterceptor.class,
             CustomEventListenerPeopleEventJmsListener.class,
+            CustomEventListenerPeopleEventJmsLoggerMetadataInterceptor.class,
 
             CustomEventListenerExampleEventEventFilter.class,
             CustomEventListenerExampleEventEventFilterInterceptor.class,
             CustomEventListenerExampleEventEventListenerInterceptorChainProvider.class,
             CustomEventListenerExampleEventEventValidationInterceptor.class,
             CustomEventListenerExampleEventJmsListener.class,
+            CustomEventListenerExampleEventJmsLoggerMetadataInterceptor.class,
 
             RecordingJsonSchemaValidator.class,
 
@@ -119,7 +123,6 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
             DefaultJmsEnvelopeSender.class,
             DefaultEnvelopeConverter.class,
             JsonSchemaValidationInterceptor.class,
-            JmsLoggerMetadataInterceptor.class,
             DefaultJmsParameterChecker.class,
             JmsAdapterToHandlerIT.TestServiceContextNameProvider.class,
             JsonSchemaLoader.class,
@@ -175,7 +178,8 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
             TestSubscriptionManagerProducer.class,
             RecordingSubscriptionManager.class,
 
-            DefaultEventSourceDefinitionFactory.class
+            DefaultEventSourceDefinitionFactory.class,
+            JmsLoggerMetadataAdder.class
     })
     public WebApp war() {
         return new WebApp()