+++ generators/rest-adapter-core/src/main/java/uk/gov/justice/services/adapter/rest/application/DefaultCommonProviders.java
@@ -2,7 +2,6 @@ package uk.gov.justice.services.adapter.rest.application;
 
 import uk.gov.justice.services.adapter.rest.cors.CorsFeature;
 import uk.gov.justice.services.adapter.rest.filter.JsonValidatorRequestFilter;
-import uk.gov.justice.services.adapter.rest.filter.LoggerRequestDataFilter;
 import uk.gov.justice.services.adapter.rest.interceptor.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.adapter.rest.mapper.BadRequestExceptionMapper;
 import uk.gov.justice.services.adapter.rest.mapper.ConflictedResourceExceptionMapper;
@@ -27,7 +26,6 @@ public class DefaultCommonProviders implements CommonProviders {
         classes.add(ConflictedResourceExceptionMapper.class);
         classes.add(ForbiddenRequestExceptionMapper.class);
         classes.add(JsonValidatorRequestFilter.class);
-        classes.add(LoggerRequestDataFilter.class);
         classes.add(JsonSchemaValidationInterceptor.class);
         classes.add(CorsFeature.class);
         return classes;